backend/requirements.txt
Flask==2.2.5
requests==2.31.0
SQLAlchemy==1.4.49
pymysql==1.1.0
pandas==2.2.3
python-dateutil==2.8.2